feat: 🎸 rename slots to slot
BREAKING CHANGE: 🧨 The property slots from widget structure is renamed to slot
