+++ packages/create-widget/views/svelte/template/src/views/viewFactory.js
@@ -2,7 +2,7 @@ import View from './View.svelte';
 import { headlineSlotFactory } from '../slots/headlineSlotFactory';
 
 async function viewFactory(widget) {
-  const slots = (await Promise.all([headlineSlotFactory(widget)])).reduce(
+  const slot = (await Promise.all([headlineSlotFactory(widget)])).reduce(
     (acc, cur) => {
       acc[cur.name] = cur;
 
@@ -13,7 +13,7 @@ async function viewFactory(widget) {
 
   return {
     View,
-    slots,
+    slot,
   };
 }
 export { viewFactory };
