+++ packages/create-widget/views/vanilla/template/src/views/View.js
@@ -1,9 +1,9 @@
 import Counter from '../components/Counter';
 import ErrorView from './ErrorView';
-import { headlineSlotFactory } from '../slots/HeadlineSlot';
+import { headlineSlotFactory } from '../slot/HeadlineSlot';
 
 async function viewFactory(widget) {
-  const slots = (await Promise.all([headlineSlotFactory(widget)])).reduce(
+  const slot = (await Promise.all([headlineSlotFactory(widget)])).reduce(
     (acc, cur) => {
       acc[cur.name] = cur;
 
@@ -14,7 +14,7 @@ async function viewFactory(widget) {
 
   return {
     View,
-    slots,
+    slot,
   };
 }
